backend/interback/requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asgiref==3.8.1
assemblyai==0.35.1
asteroid-filterbanks==0.4.0
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
av==11.0.0
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.1
ctranslate2==4.5.0
cycler==0.12.1
cymem==2.0.10
decorator==5.1.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
docopt==0.6.2
einops==0.8.0
executing==2.1.0
faster-whisper==0.10.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httplib2==0.22.0
huggingface-hub==0.26.2
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
ipython==8.29.0
jedi==0.19.2
jellyfish==1.1.0
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.9
llvmlite==0.43.0
Mako==1.3.8
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.11
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.20.1
openai-whisper==20231117
optuna==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
primePy==1.3
proglog==0.1.10
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf
proto-plus==1.25.0
psutil==6.1.0
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.4.0
pytorch-metric-learning==2.7.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
skillNer==1.0.3
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==0.5.16
SQLAlchemy==2.0.36
sqlparse==0.5.2
srsly==2.4.8
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboardX==2.6.2.2
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.15.2
torch==2.5.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.39.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==14.1
whisperx==3.1.1
wordcloud==1.9.4
wrapt==1.17.0
yarl==1.18.3
